+++ validator-constraints/src/main/java/io/avaje/validation/constraints/DecimalMin.java
@@ -16,7 +16,6 @@ import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import io.avaje.validation.Constraint;
 import io.avaje.validation.constraints.DecimalMin.List;
 
 /**
