+++ validator-constraints/src/main/java/io/avaje/validation/constraints/Null.java
@@ -11,7 +11,6 @@ import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import io.avaje.validation.Constraint;
 import io.avaje.validation.constraints.Null.List;
 
 @Constraint
