+++ validator-constraints/src/main/java/io/avaje/validation/constraints/PastOrPresent.java
@@ -11,8 +11,6 @@ import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import io.avaje.validation.Constraint;
-
 @Constraint
 @Documented
 @Target({METHOD, FIELD, ANNOTATION_TYPE, TYPE_USE})
